Enhance feedbackController security: - Added authorization checks to restrict access to authorized users only.

+++ controllers/feedbackController.js
@@ -1,5 +1,6 @@
 import { ObjectId } from "mongodb";
 import { getFeedbackCollection } from "../collections.js"
+import { authorizeUser } from "./authorizationController.js";
 
 export const getAllFeedback = async (req, res) => {
     try {
@@ -27,18 +28,24 @@ export const getFeedbackById = async (req, res) => {
         const feedbackCollection = await getFeedbackCollection();
 
         const { userId } = req.params;
-        const options = {
-            projection: {
-                _id: 0,
-                name: 1,
-                profileImage: 1,
-                headline: 1,
-                feedback: 1
-            }
-        };
 
-        const feedback = await feedbackCollection.findOne({ userId }, options);
-        res.status(200).json(feedback);
+        const authorizeStatus = await authorizeStatus(userId, req.decoded.email);
+        if (authorizeStatus === 200) {
+            const options = {
+                projection: {
+                    _id: 0,
+                    name: 1,
+                    profileImage: 1,
+                    headline: 1,
+                    feedback: 1
+                }
+            };
+
+            const feedback = await feedbackCollection.findOne({ userId }, options);
+            res.status(200).json(feedback);
+        }
+        else if (authorizeStatus === 403) res.status(403).json({ error: true, message: 'Forbidden Access' });
+
     } catch (error) {
         console.error("Error fetching feedback:", error);
         res.status(500).json({ message: "Internal server error", error: error.message });
@@ -51,15 +58,21 @@ export const addFeedback = async (req, res) => {
         const feedback = req.body;
         const userId = feedback.userId;
 
-        // Check if feedback already exists for the user
-        const isFeedbackExist = await feedbackCollection.findOne({ userId });
-        if (isFeedbackExist) {
-            return res.status(409).json({ error: true, message: 'Cannot add multiple feedbacks' });
-        }
+        const authorizeStatus = await authorizeUser(userId, req.decoded.email);
 
-        // Insert feedback into the collection
-        const result = await feedbackCollection.insertOne(feedback);
-        res.status(201).json(result);
+        if (authorizeStatus === 200) {
+            // Check if feedback already exists for the user
+            const isFeedbackExist = await feedbackCollection.findOne({ userId });
+            if (isFeedbackExist) {
+                return res.status(409).json({ error: true, message: 'Cannot add multiple feedbacks' });
+            }
+
+            // Insert feedback into the collection
+            const result = await feedbackCollection.insertOne(feedback);
+            res.status(201).json(result);
+        }
+        else if (authorizeStatus === 403) res.status(403).json({ error: true, message: 'Forbidden Access' });
+
     } catch (error) {
         console.error("Error adding feedback:", error);
         res.status(500).json({ message: "Internal server error", error: error.message });
@@ -69,20 +82,25 @@ export const addFeedback = async (req, res) => {
 export const updateFeedback = async (req, res) => {
     try {
         const feedbackCollection = await getFeedbackCollection();
-        const { userId, headline, feedback } = req.body;     //updatedFeedback only contain userId, headline and feedback             
+        const { userId, headline, feedback } = req.body;     //updatedFeedback only contain userId, headline and feedback
 
-        const filter = {
-            userId
-        };
-        const updateDoc = {
-            $set: {
-                headline,
-                feedback
-            }
-        };
+        const authorizeStatus = await authorizeUser(userId, req.decoded.email);
 
-        const result = await feedbackCollection.updateOne(filter, updateDoc, { upsert: false });
-        res.status(200).json(result);
+        if (authorizeStatus === 200) {
+            const filter = {
+                userId
+            };
+            const updateDoc = {
+                $set: {
+                    headline,
+                    feedback
+                }
+            };
+
+            const result = await feedbackCollection.updateOne(filter, updateDoc, { upsert: false });
+            res.status(200).json(result);
+        }
+        else if (authorizeStatus === 403) res.status(403).json({ error: true, message: 'Forbidden Access' });
     } catch (error) {
         console.error("Error updating feedback:", error);
         res.status(500).json({ message: "Internal server error", error: error.message });
@@ -94,8 +112,13 @@ export const removeFeedback = async (req, res) => {
         const feedbackCollection = await getFeedbackCollection();
         const { userId } = req.params;
 
-        const result = await feedbackCollection.deleteOne({ userId })
-        res.status(200).json(result);
+        const authorizeStatus = await authorizeUser(userId, req.decoded.email);
+
+        if (authorizeStatus === 200) {
+            const result = await feedbackCollection.deleteOne({ userId })
+            res.status(200).json(result);
+        }
+        else if (authorizeStatus === 403) res.status(403).json({ error: true, message: 'Forbidden Access' });
     } catch (error) {
         console.error("Error removing feedback:", error);
         res.status(500).json({ message: "Internal server error", error: error.message });
